Initial release (v0.96)
Source code for v0.96.

The unit list is incomplete and some features are not fully supported,
chiefly:

- Transform / Combine / Separate abilities. I haven't implemented the
part to recognize which units are you going to transform into / combine
with.
- MAP weapons (weapons changed to MAP only target 1 unit)
- Beam weapons. This actually works, but it's not fully known if the
not-beam weapons listed as beam (like Breast Fire) will be affected if
you change this property.

Better not to touch any of those when editing the game.